Create seirds-epidemiology-model.py
This code implements an SEIRDS compartmental model for infectious disease forecasting. SEIRDS is a variation of the SEIRS model with a compartment for deceased separate from "recovered". The variable names in these compartmental epidemiology models are very confusing # so here is a short explanation of them:
* Susceptible: People who are not exposed or infected but that are not immune, so they are susceptible to the virus
* Exposed: People who have been exposed but are not yet contagious. This variable accounts for the incubation period.
* Infected: People who have become infected by the virus and are contagious therefore able to transmit it to others.
* Recovered: People who had the virus and recovered such that they are now immune for some period
* Deceased: People who succumbed to the effects of the virus and passed away. 
* Recovered people remain immune for some period before becoming susceptible again

Note: This model currently ignores "Vital Dynamics", which accounts for people who are born or die or other causes.